move projects list to a temp hook; will replace with github repo, but that has a bunch more work, beyond this refactor
